requirements.txt
deepmerge==2.0
jsonpatch==1.33
PyYAML==5.4.1